+++ forui/lib/forui.dart
@@ -17,6 +17,7 @@ export 'widgets/checkbox.dart';
 export 'widgets/dialog.dart';
 export 'widgets/divider.dart';
 export 'widgets/header.dart';
+export 'widgets/icon.dart';
 export 'widgets/label.dart';
 export 'widgets/popover.dart';
 export 'widgets/progress.dart';