+++ forui/lib/foundation.dart
@@ -1,6 +1,6 @@
 /// Low-level utilities and services.
 library forui.foundation;
 
-export 'src/foundation/layout.dart';
+export 'src/foundation/rendering.dart' hide Alignments, RenderBoxes;
 export 'src/foundation/portal/portal.dart';
 export 'src/foundation/portal/portal_shift.dart';