+++ forui/lib/src/widgets/alert.dart
@@ -14,6 +14,8 @@ import 'package:forui/forui.dart';
 /// * [FAlertStyle] for customizing an alert's appearance.
 class FAlert extends StatelessWidget {
   /// The icon. Defaults to [FAssets.icons.circleAlert].
+  ///
+  /// [icon] is wrapped in [FIconStyle], and therefore works with [FIcon]s.
   final Widget icon;
 
   /// The title.
@@ -42,7 +44,7 @@ class FAlert extends StatelessWidget {
     this.subtitle,
     this.style = FAlertStyle.primary,
     super.key,
-  }) : icon = icon ?? FAlertIcon(icon: FAssets.icons.circleAlert);
+  }) : icon = icon ?? FIcon(FAssets.icons.circleAlert);
 
   @override
   Widget build(BuildContext context) {
@@ -61,7 +63,10 @@ class FAlert extends StatelessWidget {
           children: [
             Row(
               children: [
-                InheritedData(style: style, child: icon),
+                FInheritedIconStyle(
+                  style: FIconStyle(color: style.iconColor, size: style.iconSize),
+                  child: icon,
+                ),
                 Flexible(
                   child: Padding(
                     padding: const EdgeInsets.only(left: 8),
@@ -76,7 +81,7 @@ class FAlert extends StatelessWidget {
             if (subtitle != null)
               Row(
                 children: [
-                  SizedBox(width: style.icon.size),
+                  SizedBox(width: style.iconSize),
                   Flexible(
                     child: Padding(
                       padding: const EdgeInsets.only(top: 3, left: 8),
@@ -99,6 +104,84 @@ class FAlert extends StatelessWidget {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty('style', style));
   }
+}
+
+/// [FAlertCustomStyle]'s style.
+final class FAlertStyles with Diagnosticable {
+  /// The primary alert style.
+  final FAlertCustomStyle primary;
+
+  /// The destructive alert style.
+  final FAlertCustomStyle destructive;
+
+  /// Creates a [FAlertStyles].
+  const FAlertStyles({
+    required this.primary,
+    required this.destructive,
+  });
+
+  /// Creates a [FAlertStyles] that inherits its properties from the provided [colorScheme], [typography], and [style].
+  FAlertStyles.inherit({required FColorScheme colorScheme, required FTypography typography, required FStyle style})
+      : primary = FAlertCustomStyle(
+          padding: const EdgeInsets.fromLTRB(16, 12, 16, 12),
+          iconColor: colorScheme.foreground,
+          titleTextStyle: typography.base.copyWith(
+            fontWeight: FontWeight.w500,
+            color: colorScheme.foreground,
+            height: 1.2,
+          ),
+          subtitleTextStyle: typography.sm.copyWith(color: colorScheme.foreground),
+          decoration: BoxDecoration(
+            border: Border.all(color: colorScheme.border),
+            borderRadius: style.borderRadius,
+            color: colorScheme.background,
+          ),
+        ),
+        destructive = FAlertCustomStyle(
+          padding: const EdgeInsets.fromLTRB(16, 12, 16, 12),
+          iconColor: colorScheme.destructive,
+          titleTextStyle: typography.base.copyWith(
+            fontWeight: FontWeight.w500,
+            color: colorScheme.destructive,
+            height: 1.2,
+          ),
+          subtitleTextStyle: typography.sm.copyWith(color: colorScheme.destructive),
+          decoration: BoxDecoration(
+            border: Border.all(color: colorScheme.destructive),
+            borderRadius: style.borderRadius,
+            color: colorScheme.background,
+          ),
+        );
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('primary', primary))
+      ..add(DiagnosticsProperty('destructive', destructive));
+  }
+
+  /// Returns a copy of this [FAlertStyles] with the given properties replaced.
+  @useResult
+  FAlertStyles copyWith({
+    FAlertCustomStyle? primary,
+    FAlertCustomStyle? destructive,
+  }) =>
+      FAlertStyles(
+        primary: primary ?? this.primary,
+        destructive: destructive ?? this.destructive,
+      );
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is FAlertStyles &&
+          runtimeType == other.runtimeType &&
+          primary == other.primary &&
+          destructive == other.destructive;
+
+  @override
+  int get hashCode => primary.hashCode ^ destructive.hashCode;
 }
 
 /// A [FAlert]'s style.
@@ -131,8 +214,16 @@ final class FAlertCustomStyle extends FAlertStyle with Diagnosticable {
   /// The padding. Defaults to `const EdgeInsets.all(16)`.
   final EdgeInsets padding;
 
-  /// The icon's style.
-  final FAlertIconStyle icon;
+  /// The icon's color.
+  ///
+  /// Defaults to 20.
+  final Color iconColor;
+
+  /// The icon's size. Defaults to 20.
+  ///
+  /// ## Contract
+  /// Throws [AssertionError] if `iconSize` is not positive.
+  final double iconSize;
 
   /// The title's [TextStyle].
   final TextStyle titleTextStyle;
@@ -143,25 +234,28 @@ final class FAlertCustomStyle extends FAlertStyle with Diagnosticable {
   /// Creates a [FAlertCustomStyle].
   FAlertCustomStyle({
     required this.decoration,
-    required this.icon,
+    required this.iconColor,
     required this.titleTextStyle,
     required this.subtitleTextStyle,
     this.padding = const EdgeInsets.all(16),
-  });
+    this.iconSize = 20,
+  }) : assert(0 < iconSize, 'iconSize is $iconSize, but it should be positive.');
 
   /// Returns a copy of this [FAlertCustomStyle] with the given properties replaced.
   @useResult
   FAlertCustomStyle copyWith({
     BoxDecoration? decoration,
     EdgeInsets? padding,
-    FAlertIconStyle? icon,
+    Color? iconColor,
+    double? iconSize,
     TextStyle? titleTextStyle,
     TextStyle? subtitleTextStyle,
   }) =>
       FAlertCustomStyle(
         decoration: decoration ?? this.decoration,
         padding: padding ?? this.padding,
-        icon: icon ?? this.icon,
+        iconColor: iconColor ?? this.iconColor,
+        iconSize: iconSize ?? this.iconSize,
         titleTextStyle: titleTextStyle ?? this.titleTextStyle,
         subtitleTextStyle: subtitleTextStyle ?? this.subtitleTextStyle,
       );
@@ -172,7 +266,8 @@ final class FAlertCustomStyle extends FAlertStyle with Diagnosticable {
     properties
       ..add(DiagnosticsProperty('decoration', decoration))
       ..add(DiagnosticsProperty('padding', padding))
-      ..add(DiagnosticsProperty('icon', icon))
+      ..add(DiagnosticsProperty('iconColor', iconColor))
+      ..add(DiagnosticsProperty('iconSize', iconSize))
       ..add(DiagnosticsProperty('titleTextStyle', titleTextStyle))
       ..add(DiagnosticsProperty('subtitleTextStyle', subtitleTextStyle));
   }
@@ -184,37 +279,16 @@ final class FAlertCustomStyle extends FAlertStyle with Diagnosticable {
           runtimeType == other.runtimeType &&
           decoration == other.decoration &&
           padding == other.padding &&
-          icon == other.icon &&
+          iconColor == other.iconColor &&
+          iconSize == other.iconSize &&
           titleTextStyle == other.titleTextStyle &&
           subtitleTextStyle == other.subtitleTextStyle;
 
   @override
   int get hashCode =>
-      decoration.hashCode ^ padding.hashCode ^ icon.hashCode ^ titleTextStyle.hashCode ^ subtitleTextStyle.hashCode;
-}
-
-@internal
-class InheritedData extends InheritedWidget {
-  @useResult
-  static FAlertCustomStyle of(BuildContext context) {
-    final theme = context.dependOnInheritedWidgetOfExactType<InheritedData>();
-    return theme?.style ?? context.theme.alertStyles.primary;
-  }
-
-  final FAlertCustomStyle style;
-
-  const InheritedData({
-    required this.style,
-    required super.child,
-    super.key,
-  });
-
-  @override
-  bool updateShouldNotify(covariant InheritedData old) => style != old.style;
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(DiagnosticsProperty('style', style));
-  }
+      decoration.hashCode ^
+      padding.hashCode ^
+      iconColor.hashCode ^
+      iconSize.hashCode & titleTextStyle.hashCode ^
+      subtitleTextStyle.hashCode;
 }