+++ forui/lib/widgets/icon.dart
@@ -1,0 +1,8 @@
+/// {@category Widgets}
+///
+/// An icon that may inherit its style from its enclosing widget if supported.
+///
+/// See https://forui.dev/docs/icon for working examples.
+library forui.widgets.icon;
+
+export '../src/widgets/icon.dart';