+++ forui/test/src/widgets/alert_golden_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:forui/forui.dart';
-import 'package:forui/src/widgets/alert/alert.dart';
-import '../../test_scaffold.dart';
+import 'package:forui/src/widgets/alert.dart';
+import '../test_scaffold.dart';
 
 void main() {
   group('FAlert', () {
@@ -41,7 +41,7 @@ void main() {
               child: Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 20),
                 child: FAlert(
-                  icon: FAlertIcon(icon: FAssets.icons.badgeAlert),
+                  icon: FIcon(FAssets.icons.badgeAlert),
                   title: const Text('Alert Title'),
                   subtitle: const Text('Alert description with extra text'),
                   style: variant,