+++ samples/lib/widgets/bottom_navigation_bar.dart
@@ -34,95 +34,23 @@ class _DemoState extends State<_Demo> {
         onChange: (index) => setState(() => this.index = index),
         children: [
           FBottomNavigationBarItem(
-            icon: FAssets.icons.home,
+            icon: FIcon(FAssets.icons.home),
             label: const Text('Home'),
           ),
           FBottomNavigationBarItem(
-            icon: FAssets.icons.layoutGrid,
+            icon: FIcon(FAssets.icons.layoutGrid),
             label: const Text('Browse'),
           ),
           FBottomNavigationBarItem(
-            icon: FAssets.icons.radio,
+            icon: FIcon(FAssets.icons.radio),
             label: const Text('Radio'),
           ),
           FBottomNavigationBarItem(
-            icon: FAssets.icons.libraryBig,
+            icon: FIcon(FAssets.icons.libraryBig),
             label: const Text('Library'),
           ),
           FBottomNavigationBarItem(
-            icon: FAssets.icons.search,
-            label: const Text('Search'),
-          ),
-        ],
-      );
-}
-
-@RoutePage()
-class CustomBottomNavigationBarPage extends SampleScaffold {
-  CustomBottomNavigationBarPage({
-    @queryParam super.theme,
-  });
-
-  @override
-  Widget child(BuildContext context) => const Padding(
-        padding: EdgeInsets.all(15.0),
-        child: _CustomDemo(),
-      );
-}
-
-class _CustomDemo extends StatefulWidget {
-  const _CustomDemo();
-
-  @override
-  State<_CustomDemo> createState() => _CustomDemoState();
-}
-
-class _CustomDemoState extends State<_CustomDemo> {
-  int index = 1;
-
-  @override
-  Widget build(BuildContext context) => FBottomNavigationBar(
-        index: index,
-        onChange: (index) => setState(() => this.index = index),
-        children: [
-          FBottomNavigationBarItem.custom(
-            iconBuilder: (_, data, __) => Icon(
-              Icons.home_outlined,
-              size: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            ),
-            label: const Text('Home'),
-          ),
-          FBottomNavigationBarItem.custom(
-            iconBuilder: (_, data, __) => Icon(
-              Icons.browse_gallery_outlined,
-              size: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            ),
-            label: const Text('Browse'),
-          ),
-          FBottomNavigationBarItem.custom(
-            iconBuilder: (_, data, __) => Icon(
-              Icons.radio_outlined,
-              size: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            ),
-            label: const Text('Radio'),
-          ),
-          FBottomNavigationBarItem.custom(
-            iconBuilder: (_, data, __) => Icon(
-              Icons.library_books_outlined,
-              size: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            ),
-            label: const Text('Library'),
-          ),
-          FBottomNavigationBarItem.custom(
-            iconBuilder: (_, data, __) => Icon(
-              Icons.search_outlined,
-              size: data.itemStyle.iconSize,
-              color: data.selected ? data.itemStyle.activeIconColor : data.itemStyle.inactiveIconColor,
-            ),
+            icon: FIcon(FAssets.icons.search),
             label: const Text('Search'),
           ),
         ],