+++ forui/CHANGELOG.md
@@ -3,8 +3,6 @@
 ### Additions
 * Add `FSlider`
 
-* Add `FBottomNavigationBarItem.custom(...)`.
-
 * Add `FButtonCustomStyle.enabledHoverBoxDecoration`.
 
 * Add `FTextField.contentInsertionConfiguration`.
@@ -15,6 +13,8 @@
 
 * **Breaking** Add `FColorScheme.disabledColorBrightness` - this will only affect users that create a `FColorScheme`
   from scratch.
+
+* Add `FIcon`.
 
 ### Changes
 
@@ -36,6 +36,17 @@
 * **Breaking** Rename `FLabelStateStyle` to `FLabelStateStyles`.
 
 * **Breaking** Rename `FTextField.onSave` to `FTextField.onSaved`.
+
+* **Breaking** Remove FAlertIcon & FAlertIconStyle - use `FIcon` instead.
+
+* **Breaking** Remove FButtonIcon & FAlertIconStyle  - use `FIcon` instead.
+
+* * **Breaking** Change FButtonCustomStyle to better represent the style's layout - this will only affect users that 
+  create a custom `FButtonCustomStyle`.
+
+* **Breaking** Change `FBottomNavigationBarItem.icon` from `SvgAsset` to `Widget` - wrap the asset in ` FIcon` instead.
+
+* **Breaking** Change `FHeaderAction.icon` from `SvgAsset` to `Widget` - wrap the asset in ` FIcon` instead.
 
 ### Fixes